refactor(app): split edit screen components
